requirements.txt
# --------- pytorch --------- #
torch>=1.10.0
torchvision>=0.11.0
pytorch-lightning>=1.5.10
torchmetrics>=0.7.0

# --------- hydra --------- #
hydra-core>=1.1.0
hydra-colorlog>=1.1.0
hydra-optuna-sweeper>=1.1.0
omegaconf~=2.2.2

# --------- loggers --------- #
wandb~=0.12.21
# neptune-client
# mlflow
# comet-ml
# tensorboard

# --------- linters --------- #
pre-commit      # hooks for applying linters on commit
black           # code formatting
isort           # import sorting
flake8          # code analysis
nbstripout      # remove output from jupyter notebooks

# --------- others --------- #
python-dotenv~=0.20.0
rich~=12.5.1
pytest~=7.1.2
sh~=1.14.2
pudb            # debugger

opencv-python~=4.6.0.66
numpy~=1.22.3
pandas~=1.4.3
sklearn~=0.0
scikit-learn~=1.1.1
pillow~=9.2.0
packaging~=21.3